refactor: 구현 클래스 이동
